fix(watermark): demo拆解与规范
