+++ src/packages/watermark/doc.en-US.md
@@ -16,35 +16,7 @@ import { WaterMark } from '@nutui/nutui-react';
 
 :::demo
 
-```tsx
-import React, { useState, useRef } from "react";
-import { WaterMark, Cell, Button } from '@nutui/nutui-react';
-
-const App = () => {
-  const [flag, setFlag] = useState(false)
-  const imgSrc = useRef(
-    '//m.360buyimg.com/imagetools/jfs/t1/57345/6/20069/8019/62b995cdEd96fef03/51d3302dfeccd1d2.png'
-  )
-  return (
-    <Cell>
-      <Button onClick={() => setFlag(false)}>Text WaterMark</Button>
-      <Button onClick={() => setFlag(true)}>Image WaterMark</Button>
-      {!flag && <WaterMark zIndex={200} content="NutUI-WaterMark" />}
-      {flag && (
-        <WaterMark
-          zIndex={200}
-          content="NutUI-WaterMark"
-          rotate={22}
-          imageWidth={60}
-          imageHeight={23}
-          image={imgSrc.current}
-        />
-      )}
-    </Cell>
-  )
-}
-export default App;
-```
+<CodeBlock src='h5/demo1.tsx'></CodeBlock>
 
 :::
 
@@ -52,23 +24,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React, { useRef } from "react";
-import { WaterMark, Cell, Image } from '@nutui/nutui-react';
-
-const App = () => {
-  const src = useRef(
-    '//m.360buyimg.com/ling/jfs/t1/181258/24/10385/53029/60d04978Ef21f2d42/92baeb21f907cd24.jpg'
-  )
-  return (
-    <Cell>
-      <Image src={src.current} alt="" width="100%" height="100%" />
-      <WaterMark fullPage={false} color="red" content="nutui" />
-    </Cell>
-  )
-}
-export default App;
-```
+<CodeBlock src='h5/demo2.tsx'></CodeBlock>
 
 :::
